@@ -1,5 +1,6 @@
 CodeWorkout::Application.configure do
-  # Settings specified here will take precedence over those in config/application.rb.
+  # Settings specified here will take precedence over those in
+  # config/application.rb.
 
   # In the development environment your application's code is reloaded on
   # every request. This slows down response time but is perfect for development
@@ -13,9 +14,23 @@ CodeWorkout::Application.configure do
   config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
 
+  # For SSEs
+  config.preload_frameworks = true
+  config.allow_concurrency = true
+
   # Don't care if the mailer can't send.
-  config.action_mailer.raise_delivery_errors = false
-  config.action_mailer.default_url_options = { :host => 'localhost:3000' }
+  config.action_mailer.raise_delivery_errors = true
+  config.action_mailer.default_url_options = { :host => '192.168.33.20:3000' }
+  config.action_mailer.delivery_method = :smtp
+  config.action_mailer.smtp_settings = {
+    address: 'smtp.gmail.com',
+    port: 587,
+    domain: 'vt.edu',
+    authentication: 'plain',
+    enable_starttls_auto: true,
+    user_name: ENV['GMAIL_USERNAME'],
+    password: ENV['GMAIL_PASSWORD']
+  }
 
   # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log
@@ -27,9 +42,18 @@ CodeWorkout::Application.configure do
   # This option may cause significant delays in view rendering with a large
   # number of complex assets.
   config.assets.debug = true
+
+  # Do not fallback to assets pipeline if a precompiled asset is missed.
+  # config.assets.compile = false
+
+  # Generate digests for assets URLs.
+  # config.assets.digest = true
+
+
   config.assets.initialize_on_precompile = true
 
-#  config.log_level = :info
+  config.middleware.use LogFile::Display
+  config.log_level = :info
   config.log_formatter = proc do |severity, datetime, progname, msg|
     if severity == 'DEBUG' && msg.blank?
       ''
