@@ -1,6 +1,16 @@
-# Read about factories at https://github.com/thoughtbot/factory_girl
+# == Schema Information
+#
+# Table name: stems
+#
+#  id         :integer          not null, primary key
+#  preamble   :text(65535)
+#  created_at :datetime
+#  updated_at :datetime
+#
 
-#FactoryGirl.define do
+# Read about factories at https://github.com/thoughtbot/factory_bot
+
+#FactoryBot.define do
 #  factory :stem do
 #    preamble "MyText"
 #  end
