@@ -1,58 +1,142 @@
 class OrganizationsController < ApplicationController
-  before_action :set_organization, only: [:show, :edit, :update, :destroy]
-
-  # GET /organizations
+include AbbrHelper
+  # -------------------------------------------------------------
   def index
-    @organizations = Organization.all
+    if params[:term_id]
+      @term = Term.find(params[:term_id])
+    else
+      @term = Term.current_term
+    end
+
+    # equivalent to load_and_authorize_resource.
+    # The authorize is handled with accessible_by, then the load is
+    # performed with a custom query
+    
+    if params[:enrolled_only].to_b
+      @organizations = Organization.accessible_by(current_ability).
+        includes(courses: { course_offerings: :course_enrollments } ).
+        joins(courses: { course_offerings: :course_enrollments } ).
+        where('course_offerings.term_id' => @term, 'course_enrollments.user_id' => current_user.id).
+        distinct
+    else
+      @organizations = Organization.accessible_by(current_ability).
+        includes(courses: :course_offerings).
+        joins(courses: :course_offerings).
+        where('course_offerings.term_id' => @term).
+        distinct
+    end
   end
 
-  # GET /organizations/1
+  def search
+    if params[:slug] && params[:term]
+      @organizations = Organization.find_by(slug: params[:term]) # leaving the name pluralized for rendering
+    elsif params[:term]
+      @organizations = Organization.where('lower(name) like ? or lower(abbreviation) like ? or slug like ?',
+        "%#{params[:term].downcase}%", "%#{params[:term].downcase}%", "%#{params[:term]}%")
+    else
+      @organizations = Organization.all
+    end
+
+    render json: @organizations.to_json and return
+  end
+
+  def abbr_suggestion
+    unless params[:name]
+      render json: nil and return
+    end
+
+    name = to_title_case(params[:name])
+    strategies = ['caps', 'states', 'lowers']
+    success = false
+    abbr = nil
+
+    for strategy in strategies
+      case strategy
+      when 'caps'
+        abbr = caps_strategy(name)
+        unless (abbr && Organization.exists?(slug: abbr.downcase))
+          success = true
+          break
+        end
+      when 'states'
+        abbr = states_strategy(name)
+        unless (abbr && Organization.exists?(slug: abbr.downcase))
+          success = true
+          break
+        end
+      when 'lowers'
+        abbr = lowers_strategy(name)
+        unless (abbr && Organization.exists?(slug: abbr.downcase))
+          success = true
+          break
+        end
+      end
+    end
+
+    if abbr && success
+      render json: { abbr: abbr } and return
+    else
+      render json: { abbr: nil, msg: 'Could not auto-generate a unique abbreviation. Please enter a suitable one yourself.'} and return
+    end
+  end
+
+  # -------------------------------------------------------------
   def show
+    if params[:term_id]
+      @term = Term.find(params[:term_id])
+    else
+      @term = Term.current_term
+    end
+
+    # equivalent to load_and_authorize_resource.
+    # The authorize is handled with accessible_by, then the load is
+    # performed with a custom query
+    @organization = Organization.accessible_by(current_ability).
+      includes(courses: :course_offerings).
+      joins(courses: :course_offerings).
+      where('course_offerings.term_id' => @term).
+      find(params[:id])
   end
 
-  # GET /organizations/new
-  def new
-    @organization = Organization.new
+  def new_or_existing
+    authorize! :new_or_existing, Organization, message: 'You must be signed in to start the course setup process.'
+    render layout: 'one_column'
   end
 
-  # GET /organizations/1/edit
-  def edit
-  end
-
-  # POST /organizations
   def create
-    @organization = Organization.new(organization_params)
+    @organization = Organization.new(
+      name: to_title_case(params[:name]),
+      abbreviation: params[:abbreviation],
+      slug: params[:abbreviation].downcase,
+      is_hidden: params[:is_hidden]
+    )
 
     if @organization.save
-      redirect_to @organization, notice: 'Organization was successfully created.'
+      success = true
     else
-      render action: 'new'
+      success = false
     end
+
+    result = { success: success, id: @organization.slug }
+
+    render json: result and return
   end
 
-  # PATCH/PUT /organizations/1
-  def update
-    if @organization.update(organization_params)
-      redirect_to @organization, notice: 'Organization was successfully updated.'
-    else
-      render action: 'edit'
-    end
-  end
-
-  # DELETE /organizations/1
-  def destroy
-    @organization.destroy
-    redirect_to organizations_url, notice: 'Organization was successfully destroyed.'
-  end
-
+  #~ Private instance methods .................................................
   private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_organization
-      @organization = Organization.find(params[:id])
-    end
 
+    # -------------------------------------------------------------
     # Only allow a trusted parameter "white list" through.
     def organization_params
-      params.require(:organization).permit(:display_name, :url_part)
+      params.require(:organization).permit(:name, :abbreviation, :term_id)
     end
+
+
+    # -------------------------------------------------------------
+    # Defines resource human-readable name for use in flash messages.
+    def interpolation_options
+      { resource_name: @organization.name }
+    end
+
+
 end
