@@ -1,59 +1,22 @@
-class TermsController < ApplicationController
-  before_action :set_term, only: [:show, :edit, :update, :destroy]
+class TermsController < InheritedResources::Base
+  load_and_authorize_resource
 
-  # GET /terms
-  def index
-    @terms = Term.all
-  end
+  # All standard action methods are inherited!
 
-  # GET /terms/1
-  def show
-  end
-
-  # GET /terms/new
-  def new
-    @term = Term.new
-  end
-
-  # GET /terms/1/edit
-  def edit
-  end
-
-  # POST /terms
-  def create
-    @term = Term.new(term_params)
-
-
-    if @term.save
-      redirect_to @term, notice: 'Term was successfully created.'
-    else
-      render action: 'new'
-    end
-  end
-
-  # PATCH/PUT /terms/1
-  def update
-    if @term.update(term_params)
-      redirect_to @term, notice: 'Term was successfully updated.'
-    else
-      render action: 'edit'
-    end
-  end
-
-  # DELETE /terms/1
-  def destroy
-    @term.destroy
-    redirect_to terms_url, notice: 'Term was successfully destroyed.'
-  end
-
+  #~ Private instance methods .................................................
   private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_term
-      @term = Term.find(params[:id])
-    end
 
+    # -------------------------------------------------------------
     # Only allow a trusted parameter "white list" through.
     def term_params
       params.require(:term).permit(:season, :starts_on, :ends_on, :year)
     end
+
+
+    # -------------------------------------------------------------
+    # Defines resource human-readable name for use in flash messages.
+    def interpolation_options
+      { resource_name: @term.display_name }
+    end
+
 end
