@@ -3,15 +3,38 @@
 # Table name: terms
 #
 #  id         :integer          not null, primary key
-#  season     :integer
-#  starts_on  :date
-#  ends_on    :date
-#  year       :integer
+#  season     :integer          not null
+#  starts_on  :date             not null
+#  ends_on    :date             not null
+#  year       :integer          not null
 #  created_at :datetime
 #  updated_at :datetime
+#  slug       :string(255)      default(""), not null
+#
+# Indexes
+#
+#  index_terms_on_slug             (slug) UNIQUE
+#  index_terms_on_year_and_season  (year,season)
 #
 
+# =============================================================================
+# Represents an academic term or semester, which indicates the time period
+# for one or more course offerings.
+#
 class Term < ActiveRecord::Base
+  extend FriendlyId
+  friendly_id :display_name, use: :history
+
+  #~ Relationships ............................................................
+
+  has_many :course_offerings, inverse_of: :term, dependent: :destroy
+
+  # Orders terms in descending order (latest time first).
+  # default_scope { order('ends_on desc') }
+  default_scope { order('year desc, season desc') }
+
+
+  #~ Constants ................................................................
 
   # Hard-coded season names. It is assumed that these names contain
   # letters and spaces only -- no numbers. For example, the Summer
@@ -33,16 +56,10 @@ class Term < ActiveRecord::Base
     hash[new_key] = v
   end
 
-  # Orders terms in descending order (latest time first).
-  scope :latest_first, -> { order('year desc, season desc') }
-
 
   #~ Validation ...............................................................
 
-  validates :season, presence: true
-  validates :year, presence: true
-  validates :starts_on, presence: true
-  validates :ends_on, presence: true
+  validates_presence_of :season, :year, :starts_on, :ends_on, :slug
 
 
   #~ Class methods ............................................................
@@ -54,22 +71,18 @@ class Term < ActiveRecord::Base
 
 
   # -------------------------------------------------------------
-  def self.from_path_component(path)
-    if path =~ /([a-z]+)-(\d+)/
-      season = SEASON_PATH_NAMES[$1]
-      year = $2
-      where(season: season, year: year)
-    else
-      where('1 = 0')
-    end
+  def self.current_term
+    result = Term.
+      where('starts_on <= :now and :now <= ends_on', now: Date.today).
+      first
+    return result ? result : Term.first
   end
+
 
   #~ Instance methods .........................................................
 
   # -------------------------------------------------------------
   def contains?(date_or_time)
-    # TODO We need to make sure time zones are properly handled, probably!
-
     starts_on <= date_or_time && date_or_time < ends_on
   end
 
@@ -92,9 +105,32 @@ class Term < ActiveRecord::Base
   end
 
 
-  # -------------------------------------------------------------
-  def url_part
-    "#{SEASON_PATH_NAMES.rassoc(season).first}-#{year}"
+  # -----------------------------------------
+  def self.time_heuristic(date_string)
+    if date_string.nil?
+      puts 'INVALID Use of time_heuristic'
+      return 0
+    else
+      date_split = date_string.split("-")
+      if date_split[1]
+        date_year = date_split[0]
+        date_month = date_split[1]
+        date_day = date_split[2]
+        return date_year * 100.0 + date_month * 1.0 + date_day * 0.01
+      else
+        puts 'INVALID date_string format'
+        return 0
+      end
+    end
   end
+
+
+  #~ Private instance methods .................................................
+  private
+
+    # -------------------------------------------------------------
+    def should_generate_new_friendly_id?
+      slug.blank? || season_changed? || year_changed?
+    end
 
 end
