@@ -38,7 +38,7 @@ class UploadedRoster
     @column_hash = params[:columns] || {}
     @has_headers = params[:has_headers]
     @course_role_id = params[:course_role_id]
-    @course_offering = CourseOffering.find_by_id(params[:course_offering_id])
+    @course_offering = CourseOffering.find_by_id(params[:id])
 
     @users_created = []
     @users_enrolled = []
@@ -149,7 +149,7 @@ class UploadedRoster
           first_name = row[columns[COLUMN_FIRST_NAME]]
           last_name = row[columns[COLUMN_LAST_NAME]]
         elsif not(full_name.nil?)
-          # needs a first and last name so extract it from full name if non-existant
+          # needs a first and last name so extract it from full name
           first_name = full_name.split(" ").first
           last_name = full_name.split(" ").second
         else
