@@ -1,13 +1,15 @@
 set :stage, :staging
+set :branch, 'staging' 
+set :rails_env, 'staging'
 
 # Simple Role Syntax
 # ==================
 # Supports bulk-adding hosts to roles, the primary
 # server in each group is considered to be the first
 # unless any hosts have the primary property set.
-role :app, %w{deploy@example.com}
-role :web, %w{deploy@example.com}
-role :db,  %w{deploy@example.com}
+role :app, %w{codeworkout@codeworkoutdev.cs.vt.edu}
+role :web, %w{codeworkout@codeworkoutdev.cs.vt.edu}
+role :db,  %w{codeworkout@codeworkoutdev.cs.vt.edu}
 
 # Extended Server Syntax
 # ======================
@@ -15,7 +17,7 @@ role :db,  %w{deploy@example.com}
 # definition into the server list. The second argument
 # something that quacks like a hash can be used to set
 # extended properties on the server.
-server 'example.com', user: 'deploy', roles: %w{web app}, my_property: :my_value
+server 'codeworkoutdev.cs.vt.edu', user: 'codeworkout', roles: %w{web app db}, branch: 'staging', my_property: :my_value
 
 # you can set custom ssh options
 # it's possible to pass any option but you need to keep in mind that net/ssh understand limited list of options
@@ -26,6 +28,11 @@ server 'example.com', user: 'deploy', roles: %w{web app}, my_property: :my_value
 #    forward_agent: false,
 #    auth_methods: %w(password)
 #  }
+
+set :ssh_options, {
+  port: 2222
+}
+
 # and/or per server
 # server 'example.com',
 #   user: 'user_name',
