@@ -20,7 +20,7 @@ CodeWorkout::Application.configure do
   # config.action_dispatch.rack_cache = true
 
   # Disable Rails's static asset server (Apache or nginx will already do this).
-  config.serve_static_assets = false
+  config.serve_static_files = false
 
   # Compress JavaScripts and CSS.
   config.assets.js_compressor = :uglifier
@@ -44,6 +44,18 @@ CodeWorkout::Application.configure do
 
   # Set to :debug to see everything in the log.
   config.log_level = :info
+
+  config.log_formatter = proc do |severity, datetime, progname, msg|
+    if severity == 'DEBUG' && msg.blank?
+      ''
+    else
+      "%s [%s] %s\n" % [
+        datetime.strftime('%Y-%m-%d %H:%M:%S'),
+        severity,
+        String === msg ? msg : msg.inspect
+        ]
+    end
+  end
 
   # Prepend all log lines with the following tags.
   # config.log_tags = [ :subdomain, :uuid ]
@@ -75,21 +87,6 @@ CodeWorkout::Application.configure do
   # Disable automatic flushing of the log to improve performance.
   # config.autoflush_log = false
 
-  # Use default logging formatter so that PID and timestamp are not suppressed.
-  config.log_formatter = ::Logger::Formatter.new
-
-  config.action_mailer.default_url_options = { :host => 'http://codeworks.org/' }
-
-#  config.log_level = :warn
-  config.log_formatter = proc do |severity, datetime, progname, msg|
-    if severity == 'DEBUG' && msg.blank?
-      ''
-    else
-      "%s [%s] %s\n" % [
-        datetime.strftime('%Y-%m-%d %H:%M:%S'),
-        severity,
-        String === msg ? msg : msg.inspect
-        ]
-    end
-  end
+  config.action_mailer.default_url_options = { :host => 'http://codeworkoutdev.cs.vt.edu/' }
+  config.middleware.use LogFile::Display
 end
